moved network code in separated class
