+++ ZeldaFullEditor/Gui/ExtraForms/NetHost.cs
@@ -10,18 +10,22 @@ using System.Windows.Forms;
 
 namespace ZeldaFullEditor.Gui.ExtraForms
 {
-	public partial class NetworkForm : Form
+	public partial class NetHost : Form
 	{
-		public NetworkForm()
+		public string port;
+		public NetHost()
 		{
 			InitializeComponent();
 		}
-		public string ip = "127.0.0.1";
-		public string port = "14242";
+
 		private void button1_Click(object sender, EventArgs e)
 		{
-			this.ip = textBox1.Text;
-			this.port = textBox2.Text;
+			port = portBox.Text;
+			this.Close();
+		}
+
+		private void button2_Click(object sender, EventArgs e)
+		{
 			this.Close();
 		}
 	}
