+++ ZeldaFullEditor/Gui/ExtraForms/NetHost.resx
@@ -117,7 +117,7 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="label3.Text" xml:space="preserve">
+  <data name="label2.Text" xml:space="preserve">
     <value>This is an experimental version it may contains bugs and problems
 The overworld is fully editable in multi.
 The underworld only the tile objects are editable in multi.
